refactor : 테스트 코드 폴더 및 메소드명 리팩토링 (#98)
