+++ src/test/java/com/soma/snackexercise/controller/exgroup/ExgroupControllerAdviceTest.java
@@ -2,7 +2,7 @@ package com.soma.snackexercise.controller.exgroup;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.soma.snackexercise.advice.ExceptionAdvice;
-import com.soma.snackexercise.config.TestUserArgumentResolver;
+import com.soma.snackexercise.utils.TestUserArgumentResolver;
 import com.soma.snackexercise.dto.exgroup.request.ExgroupCreateRequest;
 import com.soma.snackexercise.dto.exgroup.request.ExgroupUpdateRequest;
 import com.soma.snackexercise.exception.*;
